requirements.txt
Django==4.1.3
ipython
Pillow
django-extensions
python-dotenv==0.21.0
lxml==4.9.1
celery~=5.2.6
redis==4.3.4
psycopg2-binary
bs4==0.0.1
beautifulsoup4~=4.11.1
requests==2.28.1
django-recaptcha
gunicorn
pytils
stateful-object~=0.1
djangorestframework~=3.14.0
djangorestframework-simplejwt

PyJWT~=2.6.0
bcrypt~=4.0.1